Move directory crawling code into a separate file_operations.js file and remove unnecessary polling function
